% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfirst_functions.R
\name{bfirst2}
\alias{bfirst2}
\title{Exhaustive breath-first search of directional network, for-loop and append version}
\usage{
bfirst2(edges, get_node, source_as_index = FALSE, nodes = NULL)
}
\arguments{
\item{edges}{directional graph represented by data.table or data.frame with columns from and to, containing node ids}

\item{get_node}{node id(s) to start from, or index into edges$from or nodes$id(if supplied) to start from}

\item{source_as_index}{bool default FALSE, is get_node the id of node or a index to input edges or nodes}

\item{nodes}{data.frame or data.table of nodes need one column of id}
}
\value{
list of all non-circular non-converged paths
}
\description{
Exhaustive breath-first search of directional network, for-loop and append version
}
\details{
same as bfirst but slightly less cloning and more appending. Uses for-loop instead of mapply/lapply. Order of  the
same returned paths differ. In bfirst ordered by first terminated. In bfirst2 by first initiated.

noddes$id, edges$from and edges$to should preferably be numeric: integer or double. Char could be slower.
The breadfirst algoritm will search through the network generating pathways. Pathway will be generated in parallel,
that all non-terminated paths will be of the same length. The algorithm computes as this:
Get latest node in all non-terminated paths. Lookup and gather lsit of edges emitting from these nodes. Drop edges that: (1)
are leading to already visited nodes in previous epocs to avoid circular paths. (2) Drop edges that converge to same node,
keep only one edge to each unique node. (3) Some nodes may implicitly have no emitting edges. Due to (1), (2) and (3) some path
will be terminated as latest node has no valid emitting edge. Terminated paths are saved for last. Non terminated paths are
extended by their respective emitting edges. If one node has e.g. two emitting edges the path will be cloned in two version
where only their new latest node differ. A new epoch starts with list of active paths. When there is no more active paths,
then all paths has been terminated. Then, the function return a list of all terminated paths.

Breadth first search provides shortest paths from source to any connected nodes.
Conversely depth first search provides (shortest and/or longest or some in between) paths from source to any connected
nodes.
See x function to any path (shortest, longest, etc.)
}
