% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppWrappers.R
\name{bfirst_test}
\alias{bfirst_test}
\title{#' Hello Hello
#'
#' @return
#' @export
#' @useDynLib familyR
#' @examples hello_hello()
hello_hello = function() {
  familyR:::rcpp_hello_world()
}
simple test}
\usage{
bfirst_test(source_node_no, from, to, n_epochs, paths_stream, seen_nodes_stream)
}
\arguments{
\item{seen_nodes_stream}{}
}
\description{
#' Hello Hello
#'
#' @return
#' @export
#' @useDynLib familyR
#' @examples hello_hello()
hello_hello = function() {
  familyR:::rcpp_hello_world()
}
simple test
}
\examples{
#blop
}
