% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_family.R
\name{copy_family}
\alias{copy_family}
\title{Copy family data}
\usage{
copy_family(graph_dt)
}
\arguments{
\item{graph_dt}{graph_dt; Graph_dt list of data.tables with edges and nodes}
}
\value{
A copied version of the inputted graph_dt data.tables.
}
\description{
Copy one list of familyR data to another.
}
\details{
As familyR relies on data.table it also builds on the principle of changing input
by reference. Therefor you need to specify, that you want to copy data from on data.table
to another insted of just creating a new reference. This is, what the copy_family does.
}
\examples{
my_network <- data.table::data.table(
  from = c("A", "A", "B", "C", "X", "Y"),
  to = c("B", "C", "D", "E", "Y", "Z")
)

my_family <- get_familytree(my_network, "B")

my_copied_family <- copy_family(my_family)

}
