% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfirst_functions.R
\name{dfirst_fastR}
\alias{dfirst_fastR}
\title{fast all paths}
\usage{
dfirst_fastR(source_node, dt, max_depth = 511, verbose = FALSE,
  only_nodes = FALSE, allocateMemory = 1e+05)
}
\arguments{
\item{allocateMemory}{}
}
\description{
fast all paths
}
