% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppWrappers.R
\name{dfirst_test}
\alias{dfirst_test}
\title{simple test}
\usage{
dfirst_test(source_node, from, to, verbose, only_nodes, paths_stream)
}
\arguments{
\item{paths_stream}{}
}
\description{
simple test
}
\examples{
#blop
}
