% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfirst_functions.R
\name{expand_persons}
\alias{expand_persons}
\title{generate all persons from a name list}
\usage{
expand_persons(namelist)
}
\arguments{
\item{namelist}{named list like list(first=c("Alice","Bob"),adj=c("Affirming","Brave"),latter=c("Allister","Buckingham"))}
}
\value{
a char vector of person names
}
\description{
generate all persons from a name list
}
