% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_children.R
\name{get_children}
\alias{get_children}
\title{Get all parrents of a node}
\usage{
get_children(graph_dt, get_node, return_nodes = TRUE)
}
\arguments{
\item{graph_dt}{data.table; A data.table with a "source" and a "target" column. See details for more.}

\item{get_node}{character; The name of the node(s) who's children you are looking for}

\item{return_nodes}{logical; If TRUE (default) the function will return both
the inputted graph_dt along with a nodes data.table. See return for more.}
}
\value{
It will return the inputtet graph_dt data.table but only with children
included.
If return_nodes is TRUE it will also retun a data.table with node ids including a
column called level showing you how many generations there are between the diffent nodes
and the inputted node (get_node).
When return_nodes is TRUE it returns both data.tables in a list called nodes and edges.
}
\description{
get_children() returns all children, grandchildren, grand grandchildren, etc. of your input node.
}
\details{
The function takes as its input a data.table (graph_dt) with at least two columns called "source" and "target" and potentially many more.
}
\examples{
my_network <- data.table::data.table(
  source = c("A", "A", "B", "C", "X", "Y", "D"),
  target = c("B", "C", "D", "E", "Y", "Z", "E")
)

my_children <- get_children(my_network, "B")

}
