% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_familytree.R
\name{get_familytree}
\alias{get_familytree}
\title{Get the entire family tree of a node}
\usage{
get_familytree(graph_dt, get_node, return_nodes = TRUE)
}
\arguments{
\item{graph_dt}{data.table; A data.table with a "source" and a "target" column. See details for more.}

\item{get_node}{character; The name of the node(s) who's family tree you want}

\item{return_nodes}{logical; If TRUE (default) the function will return both
the inputted graph_dt along with a nodes data.table. See return for more.}
}
\value{
It will return the inputted graph_dt data.table but only with family
members included. If return_nodes is TRUE it will also retun a data.table with
node ids and wrap the both in a list called nodes and edges.
}
\description{
get_familytree() returns all nodes even remotely connected to your input node.
So not only will it return all parents, grandparents, grand grandparents, etc., children, grandchildren, grand grandchildren, etc., but also the parents, grandparents and so on for those parents and children.
}
\details{
The function takes as its input a data.table (graph_dt) with at least two columns called "source" and "target" and potentially many more.
}
\examples{
my_network <- data.table::data.table(
  source = c("A", "A", "B", "C", "X", "Y"),
  target = c("B", "C", "D", "E", "Y", "Z")
)

my_family <- get_familytree(my_network, "B")

}
