% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfirst_functions.R
\name{make_nodes}
\alias{make_nodes}
\title{Generate a nodes data.frame or data.table}
\usage{
make_nodes(namelist, method = c("expand", "sample"), n = NULL,
  Class = c("data.table", "data.frame"))
}
\arguments{
\item{namelist}{named list like list(first=c("Alice","Bob"),adj=c("Affirming","Brave"),latter=c("Allister","Buckingham"))}

\item{method}{char either "expand" or "sample". Expand generates any possible combination. Sample samples combinations}

\item{n}{integer, only for method="sample", size of sample}

\item{Class}{wrap in data.frame or data.table}
}
\value{
a nodes data.table or data.frame
}
\description{
importFrom data.table data.table
}
