% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_family.R
\name{merge_family}
\alias{merge_family}
\title{Merge family data}
\usage{
merge_family(...)
}
\arguments{
\item{...}{graph_dt; Graph_dt data.tables from get_children(), get_parents() or get_familytree()}
}
\value{
merge_family() will return the inputted graph_dt data.tables
merged together in one list (containing nodes and edges).
}
\description{
Use merge_family() to merge family data into one list.
}
\examples{
my_network <- data.table::data.table(
  from = c("A", "A", "B", "C", "X", "Y"),
  to = c("B", "C", "D", "E", "Y", "Z")
)

my_parents <- get_parents(my_network, "B")
my_children <- get_children(my_network, "B")

merged_family <- merge_family(my_parents, my_children)

}
