% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_inner_relations.R
\name{remove_inner_relations}
\alias{remove_inner_relations}
\title{Remove all inner relations in a network}
\usage{
remove_inner_relations(graph_dt)
}
\arguments{
\item{graph_dt}{data.table; A data.table with a "from", a "to" and a "level" column. See details for more.}
}
\value{
It will return the inputted graph_dt data.table but with all
inner relations removed. See examples.
}
\description{
remove_inner_relations() removes all inner relations in a network - see examples for more...
}
\details{
The function takes as its input a graph_dt from either get_parents() or get_children() with at least three columns called "from", "to" and "level" potentially many more.
}
\examples{

my_network <- data.table::data.table(
  from = c("A", "B", "B", "C", "C", "E"),
  to = c("B", "C", "D", "E", "B", "A")
)

# -------------------------------
my_children <- get_children(my_network, "A")

remove_inner_relations(my_children)


# -------------------------------
my_parents <- get_parents(my_network, "E")

remove_inner_relations(my_parents)


}
