% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfirst_functions.R
\name{sample_edges}
\alias{sample_edges}
\title{generate random edges in a network}
\usage{
sample_edges(nodes, n = 5, rep = T, cluster_size = NULL,
  Class = c("data.table", "data.frame"), ...)
}
\arguments{
\item{nodes}{a data.frame or data.table with}

\item{n}{sample size}

\item{rep}{TRUE/FALSE repeated sampling?}

\item{cluster_size}{integer, subset nrow to sample from. From 1 to cluster_size rows.}

\item{Class}{return as data.table or data.frame}

\item{...}{not used}
}
\description{
generate random edges in a network
}
