% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfirst_functions.R
\name{sample_persons}
\alias{sample_persons}
\title{sample persons from a namelist}
\usage{
sample_persons(namelist, n = 5)
}
\arguments{
\item{namelist}{named list like list(first=c("Alice","Bob"),adj=c("Affirming","Brave"),latter=c("Allister","Buckingham"))}

\item{n}{number of persons to sample}
}
\value{
a char vector of person names
}
\description{
sample persons from a namelist
}
